requirements.txt
entsoe-py==0.5.2
toml==0.10.2
requests==2.27.1
gunicorn==20.1.0
pandas==1.4.2
